requirements.txt
lexica-api
markdown
latex2mathml
pygments
Flask
Gunicorn
